Base collaborator input created (two versions)
Also had to change the API around a bit to use bffApi to be able to access the collaborators.

One version of the collaborator input is just for selecting a single collaborator, and the other allows user to select multiple and displays selections as Chips.  Right now I have an issue with this component when selecting with mouse clicks.  The onSelect has a Type issue.  It's commented out currently.