Added Collaborator data from backend
It's a little confusing because the components are all referring to contributors, but they're collaborators.  Removed FC.  Created new CollaboratorData interface, as a lot of the old interface is not being used. Removed temporary collaborator dummy data.

Ash created a new ticket to rename all the Contributor stuff that should be Collaborator stuff and I'll get to that after Audio page stuff is done.