built Link content display component
based on Lance's design, it's configured right now to display a banner image of a preview of the link if it's not a link to a file, and just a file icon if the link is to a file.  I think maybe we could also just always display the bannerImage.
I feel the design could be changed overall, now that we're not fetching Open Graph data and don't have access to things like the URL's page title.