Sending coverArtUrl and subtitleUrl for audio content.
I discovered a Safari issue that seems to be well documented by others.  When passing the poster attribute to a video player, Safari only shows it before the media is played, whereas Chrome shows it continuously.  To get around this, I've added a check to see if user is running Safari and if so, I'm setting the coverArt as a backgroundImage as well.